backend/requirements/app.txt
Django==4.0.3
django-ninja==0.17.0
python-environ==0.4.54
psycopg2==2.9.3
django-bulk-update-or-create==0.3.0